Create sample_practice_lab.yaml
This playbook installs and configures the ELK stack, Kali Linux, DVWA, mod_security, and forwards DVWA logs to the ELK stack using Filebeat. It also starts the Elasticsearch, Kibana, Logstash, and Filebeat services and enables them to start automatically on boot.

You will need to modify the code to fit your specific needs and environment. For example, you may need to adjust the Elasticsearch, Kibana, and Logstash versions, and you will need to provide the templates for the Elasticsearch, Kibana, Logstash, mod_security, and Filebeat configuration files. You may also need to change the Apache log file path if it is different on your system.